Port - AllCorruptingWyrm:Makes Curse an actual curse
In the context of roleplay, Clan Tremere is not a fun experience. They
inherently lack any niche or the ability to be feared as they are in
lore. Giovanni/Cappadocians/Kiasyd can summon ghosts and Tremere can as
well, and Lasombra get more utility out of their shadow magics so with
that said, to give the Camarilla something to make people fear Clan
Tremere.

I have swapped damage types from fire (which can easily be bloodhealed I
would add) to aggravated damage. The damage might need to be scaled down
post-generation nerf but I think its a good start to let Tremere Do
Their Thing.