Port wgtjunior743- Embrace Tracking
After being transformed into a vampire/ghoul, now you can pick between "Yes" or "No" to save that change on your slot, keep in mind that WL clans for those without the WL, will be changed to their non WL versions or randomized.
For power gaming reason, if your sire is below the 13 gen, you will always be a 13th gen as well, if they are 13th gen you will be turned into a Caitiff 14th gen but unfortunately the game by default will force the 14th gen to turn 13th on the char slot.

On a side note for coders, create_disciplines does not work if you pass a list, i tried a couple of time to fix but it was not my focus so i used give_discipline directly, that also means that gargoyles created ingame don't recieve their disciplines.